+++ quickcheck-classes-base/changelog.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
-## [0.6.X.X] - 2019-XX-XX
+Note that since `quickcheck-classes` reexports larges parts of
+`quickcheck-classes-base`, changelog entries that deal with any of the
+classes from `base` are duplicated across the two changelogs.
+
+## [0.6.1.0] - 2020-09-09
 ### Added
+- Laws for `abs` and `signum`
 - Storable Set-Set Law (resolves issue 101).
 - Add laws for `quotRem` and `divMod`.
 - Use non-commutative monoid for bifoldable tests (resolves issue 98)